backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
